+++ tests/testthat/test-bundleFiles.R
@@ -133,28 +133,28 @@ test_that("ignores temporary files", {
   expect_equal(ignored, c("foo.xlsx", "foo.csv"))
 })
 
-test_that("ignores python virtual envs", {
+test_that("ignores Python virtual envs", {
   dir <- withr::local_tempdir()
-  dirCreate(file.path(dir, "test", "bin"))
-  file.create(file.path(dir, "test", "bin", "python"))
-  dirCreate(file.path(dir, "venv"))
-  file.create(file.path(dir, "venv", "somefile"))
+
+  names <- c(
+    # well-known names ...
+    ".env", ".venv", "venv",
+
+    # other names ...
+    "test"
+  )
+
+  dirCreate(file.path(dir, names, "bin"))
+  file.create(file.path(dir, names, "bin", "python"))
 
   expect_equal(bundleFiles(dir), character())
 })
 
-test_that("preserves .env when a file", {
+test_that("preserves well-known names when not Python virtual environment", {
   dir <- withr::local_tempdir()
-  file.create(file.path(dir, ".env"))
+  file.create(file.path(dir, c(".env", ".venv", "venv")))
 
-  expect_equal(bundleFiles(dir), c(".env"))
-})
-
-test_that("ignores .env when a directory", {
-  dir <- withr::local_tempdir()
-  dir.create(file.path(dir, ".env"))
-
-  expect_equal(bundleFiles(dir), character())
+  expect_setequal(bundleFiles(dir), c(".env", ".venv", "venv"))
 })
 
 # explodeFiles ------------------------------------------------------------